Put ECS into security groups
The `infrahouse/website-pod/aws` module after version 2.6 creates
dedicated security groups for the load balancer and EC2 instances.
This commit uses the 2.6 and therefore runs the ECS infrastructure in
the dedicated security groups.

The ECS host machine needs extra ECS specific rules, that are defined in
the additional security group.

In theory, the change is backward compatible, but I will bump the major
version to make migrations safer.