Facilitate migration
Well, I hit an unexpected behavior with the module dependency.
Even though website 2.6 is supposed to be backward compatible, I don't
want to apply the 2.6.* changes on all ecs modules. Yet, the 1.0 ecs module
downloads the latest 2.6.

This PR creates a short lived version 1.0.1 for sake of migration only.
It pins the website-pod version to 2.5.0.

The next ecs major version 2.* will use  website-pod 2.6+