Expose managed_termination_protection as input variable
See
https://docs.aws.amazon.com/AmazonECS/latest/developerguide/managed-instance-draining.html
for context.

By default, we want ECS to start a new container when the current is
running on an EC2 instance that is about to terminate.
If only one container is allowed to run at the same time, the second
container won't start as long as the first one is running. For instance,
Prometheus holds an exclusive lock on the data volume. In this case we
want to disable managed_termination_protection